wip: split cli store ops
